+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+<p><strong>Added</strong></p>
+<ul>
+  <li>Support dragging and dropping files into the embedded TUI.</li>
+</ul>
+
 ## [2.0.0] - 2026-06-17
 
 <p><strong>Breaking Changes</strong></p>