+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 <p><strong>Added</strong></p>
 <ul>
-  <li>Support dragging and dropping files into the embedded TUI.</li>
+  <li>Support dragging, dropping, and pasting files into the embedded TUI.</li>
 </ul>
 
 ## [2.0.0] - 2026-06-17